Added Blog Post Categories along with other improvements
This is a rather big set of changes so, hold on to your hats...

The blog page now has a category menu! This means the following:

    - The blog page now has its own layout in `_layouts` called `blog.html`
    - You can now filter blog posts according to the categor(y/ies) that
      they are placed in
    - By clicking on a category you are taken to a category page (these
      have to be added by hand - more on that later)
    - Since the category pages also need a category menu this has been
      abstracted out into `_includes/catmenu.html`. This is included in
      both `_layouts/blog.html` and `_layouts/catpage.html`
    - To create a category page simply create a HTML file with the new
      category's name in the `/blog` folder and add the following to the
      YAML front matter:
          + layout: catpage
          + category: <catname>
      where of course <catname> is the name of the category you are
      adding.
    - So that the category menu can appear on the side etc. we have
      added rules to `css/main.css`
    - Categories in the menu are sorted alphabetically and display the number
      of posts they contain

The following category pages have been added:

    - Atom
    - Emacs
    - Functional
    - Kivy
    - PyDiff
    - Python
    - Tutorial
    - Vim
    - Website

These are simply all the categories that already existed on the
website and some of them are more like tags and (in my opinion)
redundant. For example the Atom, Emacs and Vim category all simply point
to the first editor off post and nothing else. I will create an issue
referencing this PR if it gets merged where we can discuss this further.

With the introduction of the category menu there are now a number of places
where we need to list a number or posts (home, blog, categories etc.). So
to ensure we get a consistent look and feel, the for loop which iterates
through some list of posts and generates the HTML to display them has now been
but into `_includes/post_link.html`.

To include to lists of posts we now add the following to a page:

```
{% for post in <list of posts> %}
    {% include post_link.html %}
{% endfor %}
```

Where <list of posts> could be any list of posts e.g.

   - All of them `site.posts` (like on the blog page)
   - A subset of them `site.posts limit:7` (like on the home page)
   - All of them belonging to a category `site.categories[page.category]`
     (as on the new category pages)