Tutorials are now automatically listed on language pages
Following the disussion on the Christmas List (issue #25) I have
modified the language layout to automatically add tutorials to the
relevant language pages as they are posted to the blog.

There are two sections created for each language:

    - Getting Started
    - Tutorials

As discussed in #25 we thought it would be a good idea to have a
'Getting Started' track of tutorials. Guiding people from installing
compilers etc. to be able to run 'Hello World' code and get introduced
to the basics of the language.

Then the Tutorials section is for all the other tutorials related to the
language, whether they be using particular libraries, tools, language
features etc.

So for people writing tutorials this means the following:

    - In the YAML front matter of the tutorial you need to have 'tutorial'
      down as a CATEGORY.
    - In the YAML front matter of the tutorial you need to have the same
      name used for the language as in the LANGUAGE data on the
      relevant language page as a TAG
    - If the tutorial is to be a 'getting started' tutorial then you
      also need to have 'getstarted' as a TAG in the YAML front matter.

For example if I was writing a hello world tutorial for the Brainf**k
language then I would have the following front matter to my blog post

---
layout: post
title: "Hello World in Brainf**k"
categories: tutorial
tags: bf getstarted
author: Alex
comments: true
---

For people creating new langauge pages this means you would simply have
to add a 'langauge' data entry which was all _lowercase_ to represent
the langauge. For example if I was going to create a language page for
Clojure then I would have the following YAML front matter to the page

---
layout: language
title: Clojure
language: clojure
---

Finally just a bit of pedantic aesthetics I have renamed `_layouts/languages`
to `_layouts/language`