Phase 1 of css refactor
Let's face it our css was a mess, back in the beginning no one knew what
we were doing when it came to the css. So we poked and prodded until it
worked and so far so good...

However somewhere along the way we stopped using the built in support
for sass that Jekyll has. This refactor is an attempt to streamline the
css portion of the site by leveraging the power of sass. So the
following has happened:

- `main.css` is no more, this has been replaced with `main.scss`
- The majority of the contents of the old `main.css` have been scattered
  to the far winds in the many new files in the new `_sass` folder. More
  on those later.
- `main.scss` contains a number of variables which for the most part
  should allow us to make most of the cosmentic changes we need, without
  having to dive through the depths of the css.
- The variables would allow you to
  + Change the background image and colour
  + Change the fonts, font sizes and colours
  + Change the syntax highlighting schemes in `{% highlight %}` blocks
  + A few other related things
- One thing you cannot do from `main.scss` is alter the placement of
  elements on the page. Nor can you drastically change the rules applied
  to these elements apart from its colour. This is *intentional*

To change the placement of elements on the page then you need to find
the appropriate layout file in `_sass/_layouts/` and alter the rules
there.

Similarly for the individual elements of the page, if for example if you
wanted to change the overall appearance of code snippets then you would
need to edit the `_sass/_elements/_codesnippets.scss` file.

It's important to note that this refactor should be functionally
identical to the site before this. The appearance of the site should not
change this is simply one step on the way to making future changes
easier.

Finally I think it would be a good idea to elaborate on where I see this
going, that way if anyone has any improvements or suggestions then we
can incorporate them now when it should (hopefully) be easier.

- Most of the files in `_layouts` and `_includes` Should have their own
  .scss file in `_sass/_layout/` to control the css side of the layouts.
- Each major element that could appear on a page should have a .scss
  file in `_sass/_elements` to control the element's appearence accross
  the site.
- Finally we have a number of .scss files in the actual `css` folder
  that import all the snippets of css defined in the `_sass` folder that
  it needs to display it's content. Most likely we will have .scss files
  for:
  + The home page
  + Blog pages
  + The Reference section of the site
  + The Get involved section of the site

I think that covers everything, as stated above this is only phase 1
there is still a number of things to do. But `main.scss` is now
manageable and *touchwood* the site is not broken so I thought this
would be a good place to propose the changes now.

Let me know your thoughts