Adding a landing page!
All the cool websites these have a "landing page" so I thought I'd add
one to the website. A landing page is essentially an overview as to why
you should use a product/site/thing or in our case come to code club.

This essentially covers everything that was on the old home page but
hopefully it's presented a bit better. Also the redundant list of blog
posts has been removed since that's now in the archive.

Hope you like it!