Organisation and rake commands
- Theme sass files are now located in the `_sass/_themes` folder

This also adds a rake command which automates the creation of a newtheme
this means:

- It automatically creates the required data in the `_data/themes.yml`
  for the theme to be added to the dropdown in the footer
- It automatically creates a template theme file in the `css` folder
  copied from `templates/theme.scss`
- It automatically creates a folder for the user's sass files to go in
  as well as copying over an example sass file from
  `templates/themepart.scss` (this is mainly to work around git not
  adding empty folders to version control)
- It checks to see if a theme by the same name already exists - this is
  to prevent people overriding other people's themes accidentally.

I have also moved the new blog post template from a string in the
`Rakefile` to a file `templates/blogpost.md`. This way people can modify
the templates easily without having to dive into the ruby code.

Finally I have told Jekyll to ignore the templates folder and not try to
generate content from it.