Moving codesnippets to styles
If you have ever looked inside the `codesnippets.scss` file you will
have noticed it is rather complicated and for people new to css etc. it
would be rather intimidating...

To combat this there is a new `_styles` directory in the `_sass` folder
which now contains `codesnippets.scss`. The whole point of this and the
`_styles` directory in general is to provide easy to tweak styling for
various parts of the site.

The `codesnippets.scss` file basically depends on a number of `code-*`
variables which need to be defined in a theme's main `.scss` file when it
is imported. That way by setting a few colours as variables, people who
use this style can easily adjust how code snippets are displayed on the
site to their liking.

This PR also converts all existing themes to use the "new" codesnippets
style.

There may be additional styles added in the future for various aspects
of the site.