Add Twitter Timeline!
As you may or may not know, code club now has its own twitter account!
This adds the twitter feed onto the site so vistors can see it.

I really struggled on where to put this and still have it look good
accross all devices and I think what I have come up with works ok, but
please let me know your thoughts :)

Rather than display the feed at all times, the feed is 'minimised' and
the only thing that is visible is a tab at the bottom right of the
screen which says "Twitter", clicking on this will cause the feed to be
displayed with a nice entry animation. Clicking again will hide it.

Please see screenshots for more info :)

**IMPORTANT!:** While I have added the necessary CSS to the default
layout of the site so that it is positioned correctly on all themes,
it's up to the theme makers to make it look good :)

The two most important selectors are:

- `.sidetab`: This is the class of the div element which contains the
  "Twitter Tab"
- `tweets`: This is the class of the div element which contains the
  actual timeline