Add choice of light and dark twitter feeds
So it turns out that by adding the logos I discovered the perfect way to
offer a choice between light and dark twitter themes to theme makers.

By default the theme in use will be **light** this is set in the
`_sass/_layouts/_base.scss` file, so if you want the light feed in your
theme then there is no action required by you.

If however you want the dark twitter feed then, you need to add the
following CSS code to your theme:

```
.tweets > .feed > .dark {
    display: block;
}

.tweets > .feed > .light {
    display: none;
}
```