Choosing a default logo
So having both logos be displayed at once is dumb and looks stupid so
this now sets a default in the `_sass/_layouts/_base.scss` file as with
the twitter feed theme.

The default is the **blue** logo as it should be visible on most themes
we currently have. If you want the blue logo then you don't have to
change anything.

If you want to change from the default to the white logo then you need
to add the following CSS to your theme:
```
.sidetab > .blue {
    display: none;
}

.sidetab > .white {
    display: block;
}
```