Fixing twitter
This tweaks the twitter widgets so that they are actually all on screen
correctly rather than stopping halfway down.

This also fixes the cardiff red theme which was actually displaying
both timelines at once.

**IMPORTANT:** The widgets that are on the live site no longer exist so
the feed is currently broken on the live site (I deleted them before
realising it affected the site)