Document all the things!
It's finally happening the code club website is getting documented! This
adds a new collection `_docs` and the very first piece of documentation
"Project Structure". This just gives an overview as to what each folder
and file is for in the root of the repository on Github, hopefully so
that people new to the site aren't overwelmed when they see it for the
first time.