Implementing search for docs
Using lunr.js, we can now search the documentation! It isn't great just
yet since there is only two items in the documentation for now, but as
more and more it added it should become quite useful.

Although we still need to create some sort of structure as well since it
would help first time users find out what is actually documented.

In order to implement this I ended up modifying the
`_layouts/default.html` template, this adds support for a new
frontmatter term `script`

The `script` item can be used to add extra javascript scripts to the
head of an HTML document without the need for defining an entirely new
layout just for that.